backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing and ML
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4

# Data Formats
openpyxl==3.1.2
pyarrow==14.0.1
lxml==4.9.3

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Configuration and Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Monitoring and Metrics
prometheus-client==0.19.0

# File Handling
aiofiles==23.2.1

# Redis (for caching and sessions)
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: AI/ML Libraries
# openai==1.3.7  # Uncomment if using OpenAI API
# transformers==4.35.2  # Uncomment if using Hugging Face models
# torch==2.1.1  # Uncomment if using PyTorch
# tensorflow==2.15.0  # Uncomment if using TensorFlow

# Optional: Web Scraping
# beautifulsoup4==4.12.2  # Uncomment if using web scraping
# selenium==4.15.2  # Uncomment if using Selenium
# scrapy==2.11.0  # Uncomment if using Scrapy
